requirements.txt
# Web Framework
streamlit>=1.28.0

# Deep Learning
tensorflow>=2.13.0
keras>=2.13.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0

# Machine Learning
scikit-learn>=1.3.0
joblib>=1.3.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0